+++ CHANGES.md
@@ -1,9 +1,13 @@
 # Upcoming
 ## Detour
+- added BuildingSystem.AddFurniture() detour
+- added FurnitureShop.AddFurniture() detour
+- added FurnitureShop.Restock() detour
 - renamed SavableScriptableObjectDetour to FurnitureDetour
 - cleaned ReplacementLoadFromPath() up
 ## Furniture
 - added Furniture Handler -> Callback when Furniture gets loaded, which allows us to add custum scripts to the prefabs
+- added Furniture Shop Restock Handler -> sell your furnitures!
 - fixed Furniture stacking issues
 
 # v0.1.0-alpha (July 4th 2024) - Preview